refactor: dpo controller/service/repository

+++ src/app/services/sector.service.ts
@@ -2,30 +2,50 @@ import { Injectable } from '@nestjs/common';
 import { CreateSectorDto } from '../../infra/http/dtos/create-sector.dto';
 import { UpdateSectorDto } from 'src/infra/http/dtos/update-sector.dto';
 import { SectorEntity } from '../entities/sector.entity';
+import { UsersRepository } from 'src/infra/database/prisma/repositories/user.repository';
+import { NotFoundError } from '../common/errors/types/NotFoundError';
+import { SectorRepository } from '../../infra/database/prisma/repositories/sector.repository';
 
 @Injectable()
 export class SectorService {
   constructor(
-    private readonly repository: SectorService
+    private readonly repository: SectorRepository,
+    private readonly userRepository: UsersRepository
   ){}
 
-  public async create(createSectorDto: CreateSectorDto): Promise<SectorEntity> {
+  public create(createSectorDto: CreateSectorDto): Promise<SectorEntity> {
     return this.repository.create(createSectorDto);
   }
 
-  public findAll(id: string): Promise<SectorEntity[]> {
+  public async findAll(id: string): Promise<SectorEntity[]> {
+    if(!(await this.userRepository.findById(id))){
+      throw new NotFoundError('ID do usuário não encontrado!')
+    }
+
     return this.repository.findAll(id);
   }
 
-  public findOne(id: string): Promise<SectorEntity> {
+  public async findOne(id: string): Promise<SectorEntity> {
+    if(!(await this.repository.findById(id))){
+      throw new NotFoundError('ID do setor não encontrado!')
+    }
+
     return this.repository.findOne(id);
   }
 
-  public update(id: string, updateSectorDto:UpdateSectorDto): Promise<SectorEntity> {
+  public async update(id: string, updateSectorDto:UpdateSectorDto): Promise<SectorEntity> {
+    if(!(await this.repository.findById(id))){
+      throw new NotFoundError('ID do setor não encontrado!')
+    }
+
     return this.repository.update(id, updateSectorDto);
   }
 
   public async remove(id: string) {
+    if(!(await this.repository.findById(id))){
+      throw new NotFoundError('ID do setor não encontrado!')
+    }
+
     await this.repository.remove(id);
   }
 }
